# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  dto.ActivitiesResponse:
    properties:
      activities:
        items:
          $ref: '#/definitions/dto.ActivityResponse'
        type: array
      total:
        example: 10
        type: integer
    type: object
  dto.ActivityAttendeeRequest:
    type: object
  dto.ActivityCheckinRequest:
    type: object
  dto.ActivityResponse:
    properties:
      agenda:
        example: Discuss project requirements
        type: string
      created_at:
        example: "2024-01-15T08:00:00Z"
        type: string
      created_by:
        example: 1
        type: integer
      customer_id:
        example: 1
        type: integer
      end_time:
        example: "2024-01-15T12:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      location_name:
        example: Conference Room A
        type: string
      start_time:
        example: "2024-01-15T10:00:00Z"
        type: string
      status:
        example: Scheduled
        type: string
      title:
        example: Client Meeting
        type: string
      type:
        example: Meeting
        type: string
      updated_at:
        example: "2024-01-15T08:00:00Z"
        type: string
    type: object
  dto.CreateActivityRequest:
    properties:
      agenda:
        example: Discuss project requirements
        type: string
      customer_id:
        example: 1
        type: integer
      end_time:
        example: "2024-01-15T12:00:00Z"
        type: string
      location_name:
        example: Conference Room A
        type: string
      start_time:
        example: "2024-01-15T10:00:00Z"
        type: string
      status:
        example: Scheduled
        type: string
      title:
        example: Client Meeting
        type: string
      type:
        example: Meeting
        type: string
    required:
    - customer_id
    - end_time
    - start_time
    - title
    - type
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  dto.UpdateActivityRequest:
    properties:
      agenda:
        example: Updated agenda
        type: string
      end_time:
        example: "2024-01-15T12:00:00Z"
        type: string
      location_name:
        example: Conference Room B
        type: string
      start_time:
        example: "2024-01-15T10:00:00Z"
        type: string
      status:
        example: Completed
        type: string
      title:
        example: Updated Meeting
        type: string
      type:
        example: Meeting
        type: string
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: API untuk manajemen customer menggunakan Golang, Gin, dan PostgreSQL
  title: Customer Management API
  version: "1.0"
paths:
  /api/activities:
    get:
      consumes:
      - application/json
      description: Get list of all activities with pagination
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ActivitiesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all activities
      tags:
      - Activities
    post:
      consumes:
      - application/json
      description: Create a new activity for a customer
      parameters:
      - description: Activity data
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/dto.CreateActivityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ActivityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new activity
      tags:
      - Activities
  /api/activities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an activity
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete activity
      tags:
      - Activities
    get:
      consumes:
      - application/json
      description: Get a specific activity by ID
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ActivityResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get activity by ID
      tags:
      - Activities
    put:
      consumes:
      - application/json
      description: Update an existing activity
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: integer
      - description: Activity data
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ActivityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update activity
      tags:
      - Activities
  /api/activities/{id}/attendees:
    delete:
      consumes:
      - application/json
      description: Remove users as attendees from an activity
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: integer
      - description: Attendee user IDs to remove
        in: body
        name: attendees
        required: true
        schema:
          $ref: '#/definitions/dto.ActivityAttendeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove attendees from activity
      tags:
      - Activities
    post:
      consumes:
      - application/json
      description: Add users as attendees to an activity
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: integer
      - description: Attendee user IDs
        in: body
        name: attendees
        required: true
        schema:
          $ref: '#/definitions/dto.ActivityAttendeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add attendees to activity
      tags:
      - Activities
  /api/activities/{id}/checkin:
    post:
      consumes:
      - application/json
      description: Check-in to an activity with location
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: integer
      - description: Check-in data with location
        in: body
        name: checkin
        required: true
        schema:
          $ref: '#/definitions/dto.ActivityCheckinRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check-in to activity
      tags:
      - Activities
...
securityDefinitions:
  BearerAuth:
    description: Type \"Bearer\" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"